Removes duplicates
