+++ docroot/modules/custom/va_gov_post_api/src/Service/PostFacilityBase.php
@@ -8,6 +8,8 @@ use Drupal\Core\Logger\LoggerChannelFactoryInterface;
 use Drupal\Core\Messenger\MessengerInterface;
 use Drupal\Core\StringTranslation\StringTranslationTrait;
 use Drupal\post_api\Service\AddToQueue;
+use Drupal\Core\Entity\EntityInterface;
+use Drupal\node\NodeInterface;
 
 /**
  * Class PostFacilityService posts specific service info to Lighthouse.
@@ -51,6 +53,8 @@ abstract class PostFacilityBase {
    */
   protected $postQueue;
 
+  const LOVELL_TRICARE = 1039;
+
   /**
    * Constructs a new PostFacilityBase object.
    *
@@ -92,6 +96,25 @@ abstract class PostFacilityBase {
   protected function shouldDedupe() : bool {
     // If bypass_data_check setting is enabled, do not dedupe..
     return !$this->shouldBypass();
+  }
+
+  /**
+   * Checks if the entity is within the Lovell Tricare section.
+   *
+   * @param \Drupal\Core\Entity\EntityInterface $entity
+   *   Entity.
+   *
+   * @return bool
+   *   TRUE if entity is within Lovell Tricare section. FALSE otherwise.
+   */
+  protected function isLovellTricareSection(EntityInterface $entity) : bool {
+    if (($entity instanceof NodeInterface) && ($entity->hasField('field_administration'))) {
+      /** @var \Drupal\node\NodeInterface $entity*/
+      if ($entity->get('field_administration')->target_id == self::LOVELL_TRICARE) {
+        return TRUE;
+      }
+    }
+    return FALSE;
   }
 
 }
